requirements.txt
aiogram==3.27.0
chromadb==1.5.7
sentence-transformers==5.4.0
llama-cpp-python==0.3.7
python-dotenv==1.2.2
numpy==2.2.6